Untangle the priority mess. Completely rework USB integration.
The current state now boots up. When the host actually interacts with it
via USB, the call chain apparently causes a stack overflow, as we get a
HardFault on a 'ldrexb' with a crazy memory location. The smoking barrel
is 'sp' being veeery close to __sheap...

Not sure yet whether we're looking at an infinite recursion or whether
Rust is really so braindead. Stack frames all seem to be inexplicably
huge during debugging...