Solve stack smash issues (for now). Disable display, make flash operational, fix USB.
This commit removes the double buffer of the display, as apparently
the remaining ~192kB RAM are not enough for the Rust call stacks.
The stack frame for the fido dispatch function alone is 62kB(sic!).
Also disable nesting of certain IRQ combinations for now.

NVMC implementation in nrf-hal looked broken - had to be fixed.