Hook up new APDU/T1/SE050 infrastructure into our demo.
This commit retires the playground se050 implementation for a more
fully-fledged one that lives in a separate crate.

Initialization is running - note that getting the response from
"GP SELECT" takes 49 2ms polling cycles(!).

On button press of Key1, SE050 is asked to provide 16 bytes of random
data (which can be seen through delog, but are not parsed out yet).